// web/src/app/user/api/type.ts
export interface PostFromServer {
    _id: string;
    title: string;
    content: string;
    hashtags?: string[];
    createdAt: string;
    userId: UserInfo;
    images?: string[];
    videos?: string[];
    friends_tagged?: string[];
    privacy: 'public' | 'friend';
    post_link_meta?: PostLinkMeta;
    feel_count: { [key: string]: number };
    feels: Feel[];
    comments: number;
    views: number;
    my_feel?: string; // Feel của user hiện tại đang đăng nhập
}





export interface Post {
    id: string;
    title: string;
    content: string;
    createdAt: string;
    images?: string[];
    videos?: string[];
    privacy: 'public' | 'friend';
    userId: string;
    friends_tagged?: string[];
    hashtags?: string[];
    feel?: { [key: string]: string };
    feelCount?: { [key: string]: number };
    comments: number;
}


export interface EditPostModalProps {
    open: boolean;
    onClose: () => void;
    userInfo: UserInfo;
    postData?: PostFromServer;
    handleGetPost: () => void;
}
export interface UserCardProps {
    userId: string;
    name: string;
    avatar: string;
    mutualFriends?: string;
    followersCount?: number;
    isFollowing: boolean;
}



export interface ExtractLinkParams {
    url: string;
    onSuccess?: (meta: PostLinkMeta) => void;
    onError?: (msg: string) => void;
}

export interface PostLinkMeta {
    post_link_url?: string;
    post_link_title?: string;
    post_link_description?: string;
    post_link_content?: string;
     post_link_image?: string;
     post_link_video?: string;
     type?: string;
}